refactor: move Footer
