+++ src/components/Footer/index.ts
@@ -1,0 +1,2 @@
+export * from './Footer';
+export { default } from './Footer';
